Move StripeService to Payment package
