+++ src/main/java/com/example/cdr/eventsmanagementsystem/Service/Payment/PaymentService.java
@@ -16,7 +16,6 @@ import com.example.cdr.eventsmanagementsystem.DTO.Payment.RefundRequest;
 import com.example.cdr.eventsmanagementsystem.Model.User.BaseRoleEntity;
 import com.example.cdr.eventsmanagementsystem.NotificationEvent.Payment.BookingPaymentFailed;
 import com.example.cdr.eventsmanagementsystem.Service.Auth.UserSyncService;
-import com.example.cdr.eventsmanagementsystem.Service.Booking.StripeService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.ApplicationEventPublisher;
