+++ src/main/java/com/example/cdr/eventsmanagementsystem/Service/User/StatisticsManagement.java
@@ -13,7 +13,7 @@ import com.example.cdr.eventsmanagementsystem.Repository.EventRepository;
 import com.example.cdr.eventsmanagementsystem.Repository.UsersRepository.*;
 import com.example.cdr.eventsmanagementsystem.Repository.VenueBookingRepository;
 import com.example.cdr.eventsmanagementsystem.Repository.VenueRepository;
-import com.example.cdr.eventsmanagementsystem.Service.Booking.StripeService;
+import com.example.cdr.eventsmanagementsystem.Service.Payment.StripeService;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
